Move performQuery into helpers folder
